Add files via upload
The aim of this project is to develop a deep learning model
that can accurately detect and recognize human faces in an
image and can monitor whether the person in front of camera
is involved in any malpractice during the exam or not. Also to
provide a smooth interface between admin and the
user/student.

By developing a deep learning model which can automate this
process and provide users and administration with more
accurate handling of the exams held. It will not only automate
this but also there will be no chance of malpractice and
candidates appearing for exams will get equal opportunity.